+++ clang/include/clang/Sema/Initialization.h
@@ -1384,10 +1384,10 @@ public:
 
   void AddParenthesizedListInitStep(QualType T);
 
-  /// Only used when initializing array-typed structured bindings with
+  /// Only used when initializing structured bindings from an array with
   /// direct-list-initialization. Unwrap the initializer list to get the array
   /// for array copy.
-  void AddUnwrapInitListAtTheBeginning(InitListExpr *Syntactic);
+  void AddUnwrapInitListInitStep(InitListExpr *Syntactic);
 
   /// Add steps to unwrap a initializer list for a reference around a
   /// single element and rewrap it at the end.