+++ clang/docs/ReleaseNotes.rst
@@ -302,8 +302,7 @@ Bug Fixes to C++ Support
   template depth than the friend function template. (#GH98258)
 - Clang now rebuilds the template parameters of out-of-line declarations and specializations in the context
   of the current instantiation in all cases.
-- Clang now correctly handle initializing array-typed structured bindings with
-  direct-list-initialization. (#GH31813)
+- Clang now correctly handles direct-list-initialization of a structured bindings from an array. (#GH31813)
 
 Bug Fixes to AST Handling
 ^^^^^^^^^^^^^^^^^^^^^^^^^